refactor : 업체 중복 예외처리 Global에서 분리 및 ForbiddenAccessException 역할에 맞게 수정 (#590)
