+++ module-core/src/main/java/com/back2basics/infra/exception/global/ForbiddenAccessException.java
@@ -1,12 +1,11 @@
-package com.back2basics.infra.exception;
+package com.back2basics.infra.exception.global;
 
 import com.back2basics.global.response.error.CustomException;
-import com.back2basics.infra.exception.inquiry.InquiryErrorCode;
 
 public class ForbiddenAccessException extends CustomException {
 
     public ForbiddenAccessException(String message) {
-        super(message, InquiryErrorCode.INQUIRY_DELETE_NOT_ALLOWED);
+        super(message, GlobalErrorCode.FORBIDDEN_ACCESS);
     }
 
 }
