+++ module-core/src/main/java/com/back2basics/infra/exception/global/GlobalErrorCode.java
@@ -1,4 +1,4 @@
-package com.back2basics.infra.exception.inquiry;
+package com.back2basics.infra.exception.global;
 
 import com.back2basics.global.response.code.ErrorCode;
 import lombok.Getter;
@@ -7,10 +7,9 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @RequiredArgsConstructor
-public enum InquiryErrorCode implements ErrorCode {
+public enum GlobalErrorCode implements ErrorCode {
 
-    INQUIRY_NOT_FOUND(HttpStatus.NOT_FOUND, "INQ001", "문의사항을 찾을 수 없습니다."),
-    INQUIRY_FOUND_NOT_ALLOWED(HttpStatus.FORBIDDEN, "INQ002", "자신의 문의사항이 아닙니다."),
+    FORBIDDEN_ACCESS(HttpStatus.FORBIDDEN, "GEN001", "접근 권한이 없습니다."),
     ;
 
     private final HttpStatus status;
